Migration zu Discord.py v2 (#197)
* started migration to discord.py v2
- appointments, change_log, christmas, elm_street, emoji_hunt and github done.

* - learninggroup in progress

* - fixed some merge issues
- migrated roles to discordpy2

* Restructure code, remove unnecessary code and move deprecated code.

* Minor changes to appointment notifications, make /card hidden for non-mods

* Migrated support (renamed to mod_mail), news, voice, welcom and xkcd

* Migrated timer

* migrate polls

* Migrate Text Commands

* Migrate Links

* Migrate module information

* Migrate module learning groups

---------
