+++ extensions/components/module_information/scraper.py
