+++ requirements.txt
@@ -1,21 +1,20 @@
-aiohttp==3.7.4
-async-timeout==3.0.1
-attrs==20.3.0
-beautifulsoup4==4.9.3
+aiohttp==3.8.3
+aiosignal==1.3.1
+async-timeout==4.0.2
+attrs==22.2.0
+beautifulsoup4==4.11.1
 certifi==2022.12.7
-cffi==1.14.5
-chardet==3.0.4
-disnake==2.5.2
+cffi==1.15.1
+charset-normalizer==2.1.1
+discord.py==2.1.0
 emoji==1.2.0
-idna==2.10
-multidict==5.1.0
-pycparser==2.20
-PyNaCl==1.4.0
-python-dotenv==0.20.0
-requests==2.25.1
-six==1.16.0
-soupsieve==2.2.1
-tinydb==4.4.0
-typing-extensions==3.7.4.3
-urllib3==1.26.5
-yarl==1.6.3
+frozenlist==1.3.3
+idna==3.4
+multidict==6.0.4
+pycparser==2.21
+PyNaCl==1.5.0
+python-dotenv==0.21.0
+requests==2.28.2
+soupsieve==2.3.2.post1
+urllib3==1.26.14
+yarl==1.8.2
