+++ utils.py
@@ -2,8 +2,8 @@ import os
 import re
 from datetime import datetime
 
-import disnake
-from disnake import ButtonStyle
+from discord import ButtonStyle, Embed, User, Member
+from discord.ext.commands import Context
 from dotenv import load_dotenv
 
 from views.dialog_view import DialogView
@@ -16,7 +16,7 @@ async def send_dm(user, message, embed=None):
     """ Send DM to a user/member """
 
     try:
-        if type(user) is disnake.User or type(user) is disnake.Member:
+        if type(user) is User or type(user) is Member:
             if user.dm_channel is None:
                 await user.create_dm()
 
@@ -25,8 +25,11 @@ async def send_dm(user, message, embed=None):
         print(f"Cannot send DM to {user} with text: {message}")
 
 
-def is_mod(ctx):
-    author = ctx.author
+def is_mod(context_or_member):
+    if isinstance(context_or_member, Context):
+        author = context_or_member.author
+    else:
+        author = context_or_member
     roles = author.roles
 
     for role in roles:
@@ -55,7 +58,7 @@ def to_minutes(time):
 
 
 async def confirm(channel, title, description, message="", custom_prefix="", callback=None):
-    embed = disnake.Embed(title=title,
+    embed = Embed(title=title,
                           description=description,
                           color=19607)
     return await channel.send(message, embed=embed, view=DialogView([
